adjust power profile for 8996
dirty trick to get the system permissions on the
scaling_max_freq sysfs. ueventd is not getting noticed
if the cores are restored on wakeup in kernel.
So trigger one hotplug to do that. while this is
a small performance penalty if will be only once
on wakeup or screen on

also we dont need to start thermal_engine since we use
kernel thermal control anyway