OP5/T: DeviceParts: Use 1 as the HBM on value
There are no levels to the high brightness mode on this device, it is
either one or zero. The only reason two works is because the sys path
does no value validation (off is zero, on is non-zero). This will align
the HBM implementation with other ROMs so custom kernel developers don't
suffer.

Change-Id: I36705e3a2371e8caf00a51b670a9ccc24d947e51
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>